Move rust type codec to a module
