+++ sbor/src/decode.rs
@@ -1,16 +1,5 @@
-use crate::constants::*;
-use crate::rust::borrow::Cow;
-use crate::rust::borrow::ToOwned;
-use crate::rust::boxed::Box;
-use crate::rust::cell::RefCell;
-use crate::rust::collections::*;
-use crate::rust::hash::Hash;
 use crate::rust::marker::PhantomData;
-use crate::rust::mem::MaybeUninit;
-use crate::rust::ptr::copy;
-use crate::rust::rc::Rc;
 use crate::rust::string::String;
-use crate::rust::vec::Vec;
 use crate::type_id::*;
 use crate::*;
 
@@ -158,343 +147,19 @@ impl<'de, X: CustomTypeId> Decoder<'de, X> {
     }
 }
 
-impl<X: CustomTypeId> Decode<X> for () {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let value = decoder.read_byte()?;
-        match value {
-            0 => Ok(()),
-            _ => Err(DecodeError::InvalidUnit(value)),
-        }
-    }
-}
-
-impl<X: CustomTypeId> Decode<X> for bool {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let value = decoder.read_byte()?;
-        match value {
-            0 => Ok(false),
-            1 => Ok(true),
-            _ => Err(DecodeError::InvalidBool(value)),
-        }
-    }
-}
-
-impl<X: CustomTypeId> Decode<X> for i8 {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let value = decoder.read_byte()?;
-        Ok(value as i8)
-    }
-}
-
-impl<X: CustomTypeId> Decode<X> for u8 {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let value = decoder.read_byte()?;
-        Ok(value)
-    }
-}
-
-macro_rules! decode_int {
-    ($type:ident, $type_id:ident, $n:expr) => {
-        impl<X: CustomTypeId> Decode<X> for $type {
-            fn decode_with_type_id(
-                decoder: &mut Decoder<X>,
-                type_id: SborTypeId<X>,
-            ) -> Result<Self, DecodeError> {
-                decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-                let slice = decoder.read_slice($n)?;
-                let mut bytes = [0u8; $n];
-                bytes.copy_from_slice(&slice[..]);
-                Ok(<$type>::from_le_bytes(bytes))
-            }
-        }
-    };
-}
-
-decode_int!(i16, TYPE_I16, 2);
-decode_int!(i32, TYPE_I32, 4);
-decode_int!(i64, TYPE_I64, 8);
-decode_int!(i128, TYPE_I128, 16);
-decode_int!(u16, TYPE_U16, 2);
-decode_int!(u32, TYPE_U32, 4);
-decode_int!(u64, TYPE_U64, 8);
-decode_int!(u128, TYPE_U128, 16);
-
-impl<X: CustomTypeId> Decode<X> for isize {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        i64::decode_with_type_id(decoder, type_id).map(|i| i as isize)
-    }
-}
-
-impl<X: CustomTypeId> Decode<X> for usize {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        u64::decode_with_type_id(decoder, type_id).map(|i| i as usize)
-    }
-}
-
-impl<X: CustomTypeId> Decode<X> for String {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let len = decoder.read_size()?;
-        let slice = decoder.read_slice(len)?;
-        String::from_utf8(slice.to_vec()).map_err(|_| DecodeError::InvalidUtf8)
-    }
-}
-
-impl<'a, X: CustomTypeId, B: ?Sized + 'a + ToOwned<Owned = O>, O: Decode<X> + TypeId<X>> Decode<X>
-    for Cow<'a, B>
-{
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, O::type_id())?;
-        let v = O::decode_with_type_id(decoder, type_id)?;
-        Ok(Cow::Owned(v))
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>> Decode<X> for Box<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, T::type_id())?;
-        let v = T::decode_with_type_id(decoder, type_id)?;
-        Ok(Box::new(v))
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>> Decode<X> for Rc<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, T::type_id())?;
-        let v = T::decode_with_type_id(decoder, type_id)?;
-        Ok(Rc::new(v))
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>> Decode<X> for RefCell<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, T::type_id())?;
-        let v = T::decode_with_type_id(decoder, type_id)?;
-        Ok(RefCell::new(v))
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>, const N: usize> Decode<X> for [T; N] {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let element_type_id = decoder.check_type_id(T::type_id())?;
-        decoder.check_size(N)?;
-
-        // Please read:
-        // * https://doc.rust-lang.org/stable/std/mem/union.MaybeUninit.html#initializing-an-array-element-by-element
-        // * https://github.com/rust-lang/rust/issues/61956
-        //
-        // TODO: replace with `uninit_array` and `assume_array_init` once they're stable
-
-        // Create an uninitialized array
-        let mut data: [MaybeUninit<T>; N] = unsafe { MaybeUninit::uninit().assume_init() };
-
-        // Decode element by element
-        for elem in &mut data[..] {
-            elem.write(T::decode_with_type_id(decoder, element_type_id)?);
-        }
-
-        // Use &mut as an assertion of unique "ownership"
-        let ptr = &mut data as *mut _ as *mut [T; N];
-        let res = unsafe { ptr.read() };
-        core::mem::forget(data);
-
-        Ok(res)
-    }
-}
-
-macro_rules! decode_tuple {
-    ($n:tt $($idx:tt $name:ident)+) => {
-        impl<X: CustomTypeId, $($name: Decode<X>),+> Decode<X> for ($($name,)+) {
-            fn decode_with_type_id(decoder: &mut Decoder<X>, type_id: SborTypeId<X>) -> Result<Self, DecodeError> {
-                decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-                decoder.check_size($n)?;
-
-                Ok(($($name::decode(decoder)?),+))
-            }
-        }
-    };
-}
-
-decode_tuple! { 2 0 A 1 B }
-decode_tuple! { 3 0 A 1 B 2 C }
-decode_tuple! { 4 0 A 1 B 2 C 3 D }
-decode_tuple! { 5 0 A 1 B 2 C 3 D 4 E }
-decode_tuple! { 6 0 A 1 B 2 C 3 D 4 E 5 F }
-decode_tuple! { 7 0 A 1 B 2 C 3 D 4 E 5 F 6 G }
-decode_tuple! { 8 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H }
-decode_tuple! { 9 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H 8 I }
-decode_tuple! { 10 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H 8 I 9 J }
-
-impl<X: CustomTypeId, T: Decode<X>> Decode<X> for Option<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let discriminator = decoder.read_discriminator()?;
-
-        match discriminator.as_ref() {
-            OPTION_VARIANT_SOME => {
-                decoder.check_size(1)?;
-                Ok(Some(T::decode(decoder)?))
-            }
-            OPTION_VARIANT_NONE => {
-                decoder.check_size(0)?;
-                Ok(None)
-            }
-            _ => Err(DecodeError::UnknownDiscriminator(discriminator)),
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>, E: Decode<X> + TypeId<X>> Decode<X>
-    for Result<T, E>
-{
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let discriminator = decoder.read_discriminator()?;
-        match discriminator.as_ref() {
-            RESULT_VARIANT_OK => {
-                decoder.check_size(1)?;
-                Ok(Ok(T::decode(decoder)?))
-            }
-            RESULT_VARIANT_ERR => {
-                decoder.check_size(1)?;
-                Ok(Err(E::decode(decoder)?))
-            }
-            _ => Err(DecodeError::UnknownDiscriminator(discriminator)),
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X>> Decode<X> for Vec<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let element_type_id = decoder.check_type_id(T::type_id())?;
-        let len = decoder.read_size()?;
-
-        if T::type_id() == SborTypeId::U8 || T::type_id() == SborTypeId::I8 {
-            let slice = decoder.read_slice(len)?; // length is checked here
-            let mut result = Vec::<T>::with_capacity(len);
-            unsafe {
-                copy(slice.as_ptr(), result.as_mut_ptr() as *mut u8, slice.len());
-                result.set_len(slice.len());
-            }
-            Ok(result)
-        } else {
-            let mut result = Vec::<T>::with_capacity(if len <= 1024 { len } else { 1024 });
-            for _ in 0..len {
-                result.push(T::decode_with_type_id(decoder, element_type_id)?);
-            }
-            Ok(result)
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X> + Ord> Decode<X> for BTreeSet<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let elements: Vec<T> = Vec::<T>::decode_with_type_id(decoder, type_id)?;
-        Ok(elements.into_iter().collect())
-    }
-}
-
-impl<X: CustomTypeId, T: Decode<X> + TypeId<X> + Hash + Eq> Decode<X> for HashSet<T> {
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let elements: Vec<T> = Vec::<T>::decode_with_type_id(decoder, type_id)?;
-        Ok(elements.into_iter().collect())
-    }
-}
-
-impl<X: CustomTypeId, K: Decode<X> + TypeId<X> + Ord, V: Decode<X> + TypeId<X>> Decode<X>
-    for BTreeMap<K, V>
-{
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let elements = Vec::<(K, V)>::decode_with_type_id(decoder, type_id)?;
-        Ok(elements.into_iter().collect())
-    }
-}
-
-impl<X: CustomTypeId, K: Decode<X> + TypeId<X> + Hash + Eq, V: Decode<X> + TypeId<X>> Decode<X>
-    for HashMap<K, V>
-{
-    fn decode_with_type_id(
-        decoder: &mut Decoder<X>,
-        type_id: SborTypeId<X>,
-    ) -> Result<Self, DecodeError> {
-        decoder.check_preloaded_type_id(type_id, Self::type_id())?;
-        let elements: Vec<(K, V)> = Vec::<(K, V)>::decode_with_type_id(decoder, type_id)?;
-        Ok(elements.into_iter().collect())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::encode::Encode;
     use crate::encode::Encoder;
     use crate::rust::borrow::ToOwned;
+    use crate::rust::boxed::Box;
+    use crate::rust::cell::RefCell;
+    use crate::rust::collections::*;
+    use crate::rust::rc::Rc;
+    use crate::rust::string::String;
     use crate::rust::vec;
+    use crate::rust::vec::Vec;
 
     fn assert_decoding<X: CustomTypeId>(dec: &mut Decoder<X>) {
         <()>::decode(dec).unwrap();
