+++ sbor/src/encode.rs
@@ -1,13 +1,4 @@
-use crate::constants::*;
-use crate::rust::borrow::Cow;
-use crate::rust::borrow::ToOwned;
-use crate::rust::boxed::Box;
-use crate::rust::cell::RefCell;
-use crate::rust::collections::*;
-use crate::rust::hash::Hash;
 use crate::rust::marker::PhantomData;
-use crate::rust::ptr::copy;
-use crate::rust::string::String;
 use crate::rust::vec::Vec;
 use crate::type_id::*;
 
@@ -60,355 +51,12 @@ impl<'a, X: CustomTypeId> Encoder<'a, X> {
     }
 }
 
-impl<X: CustomTypeId> Encode<X> for () {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_byte(0);
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for bool {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_byte(if *self { 1u8 } else { 0u8 });
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for i8 {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_byte(*self as u8);
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for u8 {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_byte(*self);
-    }
-}
-
-macro_rules! encode_int {
-    ($type:ident, $type_id:ident) => {
-        impl<X: CustomTypeId> Encode<X> for $type {
-            #[inline]
-            fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-                encoder.write_type_id(Self::type_id());
-            }
-            #[inline]
-            fn encode_body(&self, encoder: &mut Encoder<X>) {
-                encoder.write_slice(&(*self).to_le_bytes());
-            }
-        }
-    };
-}
-
-encode_int!(i16, TYPE_I16);
-encode_int!(i32, TYPE_I32);
-encode_int!(i64, TYPE_I64);
-encode_int!(i128, TYPE_I128);
-encode_int!(u16, TYPE_U16);
-encode_int!(u32, TYPE_U32);
-encode_int!(u64, TYPE_U64);
-encode_int!(u128, TYPE_U128);
-
-impl<X: CustomTypeId> Encode<X> for isize {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        (*self as i64).encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for usize {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        (*self as u64).encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for str {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_size(self.len());
-        encoder.write_slice(self.as_bytes());
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for &str {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_size(self.len());
-        encoder.write_slice(self.as_bytes());
-    }
-}
-
-impl<X: CustomTypeId> Encode<X> for String {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.as_str().encode_body(encoder);
-    }
-}
-
-impl<'a, X: CustomTypeId, B: ?Sized + 'a + ToOwned + Encode<X> + TypeId<X>> Encode<X>
-    for Cow<'a, B>
-{
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(B::type_id())
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.as_ref().encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for Box<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(T::type_id())
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.as_ref().encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for RefCell<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(T::type_id())
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.borrow().encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for [T] {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(T::type_id());
-        encoder.write_size(self.len());
-        if T::type_id() == SborTypeId::U8 || T::type_id() == SborTypeId::I8 {
-            let mut buf = Vec::<u8>::with_capacity(self.len());
-            unsafe {
-                copy(self.as_ptr() as *mut u8, buf.as_mut_ptr(), self.len());
-                buf.set_len(self.len());
-            }
-            encoder.write_slice(&buf);
-        } else {
-            for v in self {
-                v.encode_body(encoder);
-            }
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>, const N: usize> Encode<X> for [T; N] {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.as_slice().encode_body(encoder);
-    }
-}
-
-macro_rules! encode_tuple {
-    ($n:tt $($idx:tt $name:ident)+) => {
-        impl<X: CustomTypeId, $($name: Encode<X>),+> Encode<X> for ($($name,)+) {
-            #[inline]
-            fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-                encoder.write_type_id(Self::type_id());
-            }
-            #[inline]
-            fn encode_body(&self, encoder: &mut Encoder<X>) {
-                encoder.write_size($n);
-
-                $(self.$idx.encode(encoder);)+
-            }
-        }
-    };
-}
-
-encode_tuple! { 2 0 A 1 B }
-encode_tuple! { 3 0 A 1 B 2 C }
-encode_tuple! { 4 0 A 1 B 2 C 3 D }
-encode_tuple! { 5 0 A 1 B 2 C 3 D 4 E }
-encode_tuple! { 6 0 A 1 B 2 C 3 D 4 E 5 F }
-encode_tuple! { 7 0 A 1 B 2 C 3 D 4 E 5 F 6 G }
-encode_tuple! { 8 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H }
-encode_tuple! { 9 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H 8 I }
-encode_tuple! { 10 0 A 1 B 2 C 3 D 4 E 5 F 6 G 7 H 8 I 9 J }
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for Option<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        match self {
-            Some(v) => {
-                encoder.write_discriminator(OPTION_VARIANT_SOME);
-                encoder.write_size(1);
-                v.encode(encoder);
-            }
-            None => {
-                encoder.write_discriminator(OPTION_VARIANT_NONE);
-                encoder.write_size(0);
-            }
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X>, E: Encode<X>> Encode<X> for Result<T, E> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        match self {
-            Ok(o) => {
-                encoder.write_discriminator(RESULT_VARIANT_OK);
-                encoder.write_size(1);
-                o.encode(encoder);
-            }
-            Err(e) => {
-                encoder.write_discriminator(RESULT_VARIANT_ERR);
-                encoder.write_size(1);
-                e.encode(encoder);
-            }
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for Vec<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        self.as_slice().encode_body(encoder);
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X>> Encode<X> for BTreeSet<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(T::type_id());
-        encoder.write_size(self.len());
-        for v in self {
-            v.encode_body(encoder);
-        }
-    }
-}
-
-impl<X: CustomTypeId, T: Encode<X> + TypeId<X> + Ord + Hash> Encode<X> for HashSet<T> {
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(T::type_id());
-        encoder.write_size(self.len());
-        for v in self {
-            v.encode_body(encoder);
-        }
-    }
-}
-
-impl<X: CustomTypeId, K: Encode<X> + TypeId<X>, V: Encode<X> + TypeId<X>> Encode<X>
-    for BTreeMap<K, V>
-{
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(<(K, V)>::type_id());
-        encoder.write_size(self.len());
-        for (k, v) in self {
-            encoder.write_size(2);
-            k.encode(encoder);
-            v.encode(encoder);
-        }
-    }
-}
-
-impl<X: CustomTypeId, K: Encode<X> + TypeId<X> + Ord + Hash, V: Encode<X> + TypeId<X>> Encode<X>
-    for HashMap<K, V>
-{
-    #[inline]
-    fn encode_type_id(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(Self::type_id());
-    }
-    #[inline]
-    fn encode_body(&self, encoder: &mut Encoder<X>) {
-        encoder.write_type_id(<(K, V)>::type_id());
-        encoder.write_size(self.len());
-        let keys: BTreeSet<&K> = self.keys().collect();
-        for key in keys {
-            encoder.write_size(2);
-            key.encode(encoder);
-            self.get(key).unwrap().encode(encoder);
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::rust::borrow::ToOwned;
+    use crate::rust::boxed::Box;
+    use crate::rust::collections::*;
     use crate::rust::vec;
     use crate::NoCustomTypeId;
